requirements.txt
tiktoken
requests
python-dotenv